Extracted timezone computation into helper, added useMemo hook
